Move geometry functions to their own Geometry.h/cpp
