+++ gemrb/includes/globals.h
@@ -156,11 +156,7 @@ enum GameFeatureFlags : uint32_t {
 class Scriptable;
 class Actor;
 
-GEM_EXPORT double AngleFromPoints(const Point& p1, const Point& p2);
-GEM_EXPORT Point RotatePoint(const Point& p, double angle);
-GEM_EXPORT unsigned int Distance(const Point &pos, const Point &pos2);
 GEM_EXPORT unsigned int Distance(const Point &pos, const Scriptable *b);
-GEM_EXPORT unsigned int SquaredDistance(const Point &pos, const Point &pos2);
 GEM_EXPORT unsigned int SquaredMapDistance(const Point &pos, const Scriptable *b);
 GEM_EXPORT unsigned int PersonalDistance(const Point &p, const Scriptable *b);
 GEM_EXPORT unsigned int SquaredPersonalDistance(const Point &pos, const Scriptable *b);
